requirements.txt
langchain
langgraph
langchain-openai
pydantic
python-dotenv
duckduckgo-search
langchain-community
pydantic-settings
tenacity
